AB#171174 - renamed namespace
