+++ tests/UKHO.S100PermitService.API.FunctionalTests/Configuration/PermitServiceApiConfiguration.cs
@@ -1,4 +1,4 @@
-﻿namespace UKHO.S100PermitService.API.FunctionalTests.Helpers
+﻿namespace UKHO.S100PermitService.API.FunctionalTests.Configuration
 {
     public class PermitServiceApiConfiguration
     {
